Move Command into new command.rs
src/structs/mod.rs was the largest file (347 LOC) and this was a low hanging fruit to split it
