+++ src/slash/argument/slash_macro.rs
@@ -2,7 +2,6 @@
 
 #[allow(unused_imports)] // import is required if serenity simdjson feature is enabled
 use crate::serenity::json::prelude::*;
-use crate::serenity_prelude as serenity;
 
 /// Possible errors when parsing slash command arguments
 #[derive(Debug)]
